Move methods to internal + tight type coupling
